Refactor duplicated utilities into shared modules
