+++ shared/__init__.py
@@ -2,6 +2,13 @@
 
 Modules:
     - dataset_loader: DatasetLoader
+    - infer_file_features: infer_file_features
+    - initialize_or_load_metadata: initialize_or_load_metadata
+    - list_matching_files: list_matching_files
     - logging: get_logger
+    - read_metadata: read_metadata
     - save_dataset: save_dataset
+    - save_metadata: save_metadata
+    - write_metadata: write_metadata
+
 """
